requirements.txt
streamlit==1.32.2
pandas==2.1.4
paramiko==3.4.0
Pillow==10.1.0
python-dotenv==1.0.0
